Separate builtin function registration from the main function
