+++ include/builtin_func.h
@@ -1,0 +1,6 @@
+#ifndef BUILTIN_FUNC_H
+#define BUILTIN_FUNC_H
+
+void carrot_register_all_builtin_func(Interpreter *interpreter);
+
+#endif
